\name{HMMcopy internal functions}
\alias{HMMcopy-internal}
\alias{normalize}
\title{HMMcopy internal functions}
\description{HMMcopy internal functions for HMMsegment}
\usage{
  normalize(A)
  dirichletpdf(x, alpha)
  tdistPDF(x, mu, lambda, nu)
  estimateTNoiseParamsMap(y, mu, lambda, nu, rho, eta, m, gamma, S, kappa)
}
\details{Not for users}
\keyword{internal}
